% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_pipeline.R
\name{is_discovery_pipeline}
\alias{is_discovery_pipeline}
\title{Test whether an object is a Discovery Pipeline}
\usage{
is_discovery_pipeline(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"discovery_pipeline"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a Discovery Pipeline
}
