% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{print.benchmark_evidence}
\alias{print.benchmark_evidence}
\title{Print Benchmark Evidence}
\usage{
\method{print}{benchmark_evidence}(x, ...)
}
\arguments{
\item{x}{A \code{benchmark_evidence} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Benchmark Evidence
}
