% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{print.benchmark_object}
\alias{print.benchmark_object}
\title{Print a Benchmark Object}
\usage{
\method{print}{benchmark_object}(x, ...)
}
\arguments{
\item{x}{A \code{benchmark_object} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Benchmark Object
}
