% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_output.R
\name{print.clinical_output_object}
\alias{print.clinical_output_object}
\title{Print a Clinical Output Object}
\usage{
\method{print}{clinical_output_object}(x, ...)
}
\arguments{
\item{x}{A \code{clinical_output_object} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Clinical Output Object
}
