% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_translation.R
\name{print.clinical_translation_object}
\alias{print.clinical_translation_object}
\title{Print a Clinical Translation Object}
\usage{
\method{print}{clinical_translation_object}(x, ...)
}
\arguments{
\item{x}{A \code{clinical_translation_object} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Clinical Translation Object
}
