% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_output.R
\name{print.clinical_translation_package}
\alias{print.clinical_translation_package}
\title{Print a Clinical Translation Package}
\usage{
\method{print}{clinical_translation_package}(x, ...)
}
\arguments{
\item{x}{A \code{clinical_translation_package} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Clinical Translation Package
}
