% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_pipeline.R
\name{print.discovery_pipeline}
\alias{print.discovery_pipeline}
\title{Print a Discovery Pipeline}
\usage{
\method{print}{discovery_pipeline}(x, ...)
}
\arguments{
\item{x}{A \code{discovery_pipeline} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Discovery Pipeline
}
