% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence_integration.R
\name{print.evidence_integration_object}
\alias{print.evidence_integration_object}
\title{Print an Evidence Integration Object}
\usage{
\method{print}{evidence_integration_object}(x, ...)
}
\arguments{
\item{x}{An \code{evidence_integration_object} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Evidence Integration Object
}
