% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence_integration.R
\name{print.interpretation_package}
\alias{print.interpretation_package}
\title{Print an Interpretation Package}
\usage{
\method{print}{interpretation_package}(x, ...)
}
\arguments{
\item{x}{An \code{interpretation_package} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Interpretation Package
}
