% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{print.model_candidate}
\alias{print.model_candidate}
\title{Print a Model Candidate}
\usage{
\method{print}{model_candidate}(x, ...)
}
\arguments{
\item{x}{A \code{model_candidate} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Model Candidate
}
