% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{print.modeling_evidence}
\alias{print.modeling_evidence}
\title{Print Modeling Evidence}
\usage{
\method{print}{modeling_evidence}(x, ...)
}
\arguments{
\item{x}{A \code{modeling_evidence} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Modeling Evidence
}
